Add unit tests and re-structure
